+++ src/EventSourcingTests/Aggregation/stream_compacting.cs
@@ -361,7 +361,7 @@ public class Letters : IRevisioned
     public int BCount { get; set; }
     public int CCount { get; set; }
     public int DCount { get; set; }
-    public int Version { get; set; }
+    public long Version { get; set; }
 }
 
 public class LetterCounts: IRevisioned
@@ -371,7 +371,7 @@ public class LetterCounts: IRevisioned
     public int BCount { get; set; }
     public int CCount { get; set; }
     public int DCount { get; set; }
-    public int Version { get; set; }
+    public long Version { get; set; }
 }
 
 public class LetterCountsByString: IRevisioned
@@ -381,7 +381,7 @@ public class LetterCountsByString: IRevisioned
     public int BCount { get; set; }
     public int CCount { get; set; }
     public int DCount { get; set; }
-    public int Version { get; set; }
+    public long Version { get; set; }
 }
 
 public class LetterCountsByStringProjection: SingleStreamProjection<LetterCountsByString, string>
